feat: adjust test and classes for new response
